requirements.txt
pypdf
langchain
langchain-chroma
# chromadb # Vector storage
pytest
boto3
langchain-groq
python-dotenv
langchain_community
markdown